+++ dags/rock_content_item_categories.py
@@ -1,22 +1,8 @@
 from airflow.models import Variable
-from datetime import datetime, timedelta
 from airflow.hooks.postgres_hook import PostgresHook
-from apollos_type import apollos_id
-
+from utilities import safeget
 
 import requests
-
-
-def safeget(dct, *keys):
-    for key in keys:
-        try:
-            dct = dct[key]
-        except KeyError:
-            return None
-        except TypeError:
-            return None
-    return dct
-
 
 def fetch_and_save_content_item_categories(ds, *args, **kwargs):
     if 'client' not in kwargs or kwargs['client'] is None:
@@ -141,10 +127,12 @@ def attach_content_item_categories(ds, *args, **kwargs):
         print(params)
 
         r = requests.get(
-                f"{Variable.get(kwargs['client'] + '_rock_api')}/ContentItems",
+                f"{Variable.get(kwargs['client'] + '_rock_api')}/ContentChannelItems",
                 params=params,
                 headers=headers)
         rock_objects = r.json()
+
+        print(rock_objects);
 
         if not isinstance(rock_objects, list):
             print(rock_objects)
@@ -158,11 +146,11 @@ def attach_content_item_categories(ds, *args, **kwargs):
         skip += top
         fetched_all = len(rock_objects) < top
 
-        def update_content(obj):
-            pg_hook.run("""
+        def update_content_queries(obj):
+            return """
             UPDATE "contentItems"
-            SET "contentCategoryId" = (SELECT id FROM "contentItemCategories" WHERE "originId" = {})
-            WHERE "originId" = {};
-            """.format(str(safeGet(obj, 'ContentChannel', 'Id')), str(obj['Id'])))
+            SET "contentItemCategoryId" = (SELECT id FROM "contentItemCategories" WHERE "originId" = '{}')
+            WHERE "originId" = '{}';
+            """.format(str(safeget(obj, 'ContentChannel', 'Id')), str(obj['Id']))
 
-        map(update_content, rock_objects)
+        pg_hook.run(list(map(update_content_queries, rock_objects)))
