+++ dags/utilities.py
@@ -1,0 +1,9 @@
+def safeget(dct, *keys):
+    for key in keys:
+        try:
+            dct = dct[key]
+        except KeyError:
+            return None
+        except TypeError:
+            return None
+    return dct
